Rename a file to prevent conflicts
