+++ src/CardinalPlugin.cpp
@@ -32,7 +32,7 @@
 
 #include "CardinalCommon.hpp"
 #include "DistrhoPluginUtils.hpp"
-#include "PluginContext.hpp"
+#include "CardinalPluginContext.hpp"
 #include "extra/Base64.hpp"
 #include "extra/ScopedDenormalDisable.hpp"
 #include "extra/ScopedSafeLocale.hpp"
