+++ src/CardinalPluginContext.hpp
