+++ src/CardinalRemote/RemoteUI.hpp
@@ -1,6 +1,6 @@
 /*
  * DISTRHO Cardinal Plugin
- * Copyright (C) 2021-2022 Filipe Coelho <falktx@falktx.com>
+ * Copyright (C) 2021-2024 Filipe Coelho <falktx@falktx.com>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "NanoVG.hpp"
-#include "PluginContext.hpp"
+#include "CardinalPluginContext.hpp"
 #include "WindowParameters.hpp"
 
 #include <widget/Widget.hpp>
