+++ src/custom/glfw.cpp
@@ -1,6 +1,6 @@
 /*
  * DISTRHO Cardinal Plugin
- * Copyright (C) 2021-2023 Filipe Coelho <falktx@falktx.com>
+ * Copyright (C) 2021-2024 Filipe Coelho <falktx@falktx.com>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -16,7 +16,7 @@
  */
 
 #include "Application.hpp"
-#include "PluginContext.hpp"
+#include "CardinalPluginContext.hpp"
 
 #include <GLFW/glfw3.h>
