+++ src/override/MenuBar.cpp
@@ -1,6 +1,6 @@
 /*
  * DISTRHO Cardinal Plugin
- * Copyright (C) 2021-2023 Filipe Coelho <falktx@falktx.com>
+ * Copyright (C) 2021-2024 Filipe Coelho <falktx@falktx.com>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -53,7 +53,7 @@
 
 #include "../CardinalCommon.hpp"
 #include "../CardinalRemote.hpp"
-#include "../PluginContext.hpp"
+#include "../CardinalPluginContext.hpp"
 #include "DistrhoPlugin.hpp"
 #include "DistrhoStandaloneUtils.hpp"
