+++ src/override/Window.cpp
@@ -1,6 +1,6 @@
 /*
  * DISTRHO Cardinal Plugin
- * Copyright (C) 2021-2023 Filipe Coelho <falktx@falktx.com>
+ * Copyright (C) 2021-2024 Filipe Coelho <falktx@falktx.com>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -46,7 +46,7 @@
 #include "Application.hpp"
 #include "extra/String.hpp"
 #include "../CardinalCommon.hpp"
-#include "../PluginContext.hpp"
+#include "../CardinalPluginContext.hpp"
 #include "../WindowParameters.hpp"
 
 #ifndef DGL_NO_SHARED_RESOURCES
